requirements.txt
flask>=2.0.0
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.38.0
